pyproject.toml
[project]
name = "actividad-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "chromadb>=1.0.9",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.60",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.5",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
